refactor(wise_feedback): address review on the modern form UI
Split feedback_form.dart (432 lines) so each widget owns a file, per
review. The private _Header, _CircleButton and _LabeledField classes and
the _input/_dropdown builder methods move to lib/src/screens/widgets/;
the grabber and the inline error row become widgets too, which takes the
form itself down to 197 lines. Dart privacy is per library, so the classes
lose their underscore — they stay out of the public API because
screens.dart does not export the new directory.

Go directional on every inset with a horizontal edge:
EdgeInsetsDirectional.fromSTEB for the header and field padding,
EdgeInsetsDirectional.only(start:) for the field label, and
AlignmentDirectional.centerStart for the field box. Vertical-only and
symmetric insets are left alone — they have no handedness.

Drop letterSpacing: -0.43 from the input style. The value came off a
design file calibrated to its own typeface, but the package sets no
fontFamily, so it applied negative tracking to whatever font the platform
supplied.

Use a tear-off for onClose, and delete the comment narrating the scroll
view's structure.
